Bump axum to 0.7.4.
The Rust ecosystem is progressing rapidly. Let's try to stay up to date
as much as we can, to avoid running into a situation where we are so
outdated that it becomes non-trivial to update.

As of axum 0.7.0, axum::Server was removed, as it was removed from hyper
1.0. Use axum::serve instead. This requires bumping a few more
dependencies to satisfy required traits.