Initial TOML config file support
Add initial TOML config file support. This is in preparation of upcoming
multi-backend support. The only keys that are currently used when
starting the daemon, are the following keys under the daemon table:

- instance_id
- listen_ip
- listen_port
- otlp_endpoint

If instance_id is not set, a random UUID v4 is generated. The
instance_id key is used to set OTLP service.instance.id.

The test tries to deserialize the ai-router.toml file. To run it:

```
cargo test -- --show-output
```

If it succeeds, it will print a JSON representation of the TOML file for
comparison.

Struct members defined as Option<Type> makes it possible to omit those
fields from the config file.