dockerignore: remove Cargo.lock
Historically it was recommended to commit Cargo.lock for packages, but
not libraries. This is no longer the case. Instead, it's recommended to
do what is best for the project. While I personally prefer not having
Cargo.lock under version control, as it results in constant git
merge/rebase/cherry-pick conflicts, it does make sure the committed code
actually builds.

Due to the recent openai_dive 0.4.4 release, if we ignore Cargo.lock,
the code currently no longer builds. The fix is relatively easy, but
it's already fixed in a current feature branch. Unfortunately that
commit cannot easily be cherry-picked, due to too many other changes in
that feature branch.

Let's remove Cargo.lock from .dockerignore for now, to fix the Docker
build. As this project evolves, we can reconsider keeping Cargo.lock in
version control entirely, but we should probably have some CI in place
before we do so.

See https://blog.rust-lang.org/2023/08/29/committing-lockfiles.html.