benches: introduce benchmark support with Criterion.rs
One of the known limitations of Criterion.rs is that it is not possible
to benchmark non-public functions, so in this commit we make the
transform_triton_f32_array and build_embedding_response_data functions
public. This causes Clippy warnings. The warning for
transform_triton_f32_array can be silenced by adding the `#[must_use]`
attribute, and the warning for build_embedding_response_data can be
silenced by documenting possible errors for the function.