backend/triton: don't assume output is at index 0
Depending on the Triton config, the requested output might not be at
index 0 in the response array. Add a helper function to find the correct
index.

If getting the index fails, for streaming requests, we yield a
server-sent event, without event type, with an error in the event data
field. This way, the OpenAI python client raises an exception:

openai.APIError: An error occurred during streaming

Without this, the OpenAI chat completion client would fail in a less
descriptive way:

    first_response_time = output_start_time - start_time
                          ~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~
TypeError: unsupported operand type(s) for -: 'NoneType' and 'float'

Ideally we would return an internal server error, but it seems
prohibitively difficult to do so.