treewide: initialize HF tokenizers during startup
If we want to support max_input, we need to be able to calculate the
number of tokens for the request. Initialize HF tokenizers for each
unique occurence of hf_model_name in the models in the config file.

We cannot use the model name from the request, as this would make
max_input incompatible for models where we map an arbitrary name to a
different backend models.

Using the backend model name would require Triton to be set up using the
exact HF model name, which is not very flexible.