backend/openai: improve error handling
Try to transform errors received from the openai_dive client into
AiRouterError instead of using unwrap.

Unfortunately there are no real docs for errors returned by the OpenAI
API. As a result, there are a lot of inconsistencies between different
implementations, making it hard to properly handle all of them.

OpenAI returns a string or null for the code and param fields.
Anyscale returns a numeric value for code and an empty dict for param.

Fortunately we can use DefaultOnError, so instead of failing to
deserialize the Anyscale error, those two Option fields would be set to
None.

Not particularly happy with this code, but it makes it possible to
replace a few more unwrap and send a somewhat helpful, OpenAI
compatible error to the client.