triton/completions: handle max_tokens as for chat
Having max_tokens default to 16 is problematic for clients that don't
support setting max_tokens. At the same time Nvidia Triton requires
max_tokens to be set, so let's do the same as we did for chat
completions: use a hardcoded default that is likely high enough to
support all models.

This could still be problematic if we send a prompt that would generate
a response larger than the model limit, as we've seen during testing
with Mistral-7B-Instruct-v0.2:

[TensorRT-LLM][WARNING] Number of requested output tokens (131072) exceeds maximum sequence length (18432). Number of requested output tokens is changed t
o (17905).

While at it, change the field type to u32, to be consistent with
openai_dive's ChatCompletionParameters struct.

Ref: 11d3d43f3cee ("triton/chat: always add max_tokens").