config: add max_tokens to AiRouterModel
As explained in the previous commit, the hardcoded value of 128k for
max_tokens could still cause problems. Add a max_tokens field to the
AiRouterModel struct, and use it in chat completions and legacy
completions to a Triton backend.

If max_input is set in the OpenAI API request, we use that, value. If
not, we fall back to what is set in the model config, and if that is not
set, we use the hardcoded default.