telemetry: use same log format for debug and release builds
Having JSON log format in release build and somewhat human-readable log
format in debug builds is confusing. Let's use the same format for both.

The downside is that we can't flatten the events if we're not using
JSON, but the JSON log is really difficult to read, and since this is
what's written to stdout, let's just keep it readable.