audio: add transcriptions to Triton backend
This currently only supports WAV files with a max audio length of 30
seconds, as that is what Whisper in Triton currently supports. It also
expects the file to be single channel and 16kHz bit rate. If the
uploaded file has multiple channels, or uses a different bit rate,
ai-router will downmix and transcode the file as needed. This works up
to 8-channel audio files, as that is what the fon crate supports.