config: add option to increase max HTTP body size
The default HTTP request body limit in axum is 2MB. When a request body
exceeds this limit, there is no clear error message. E.g. with the
recently added Audio Transcription support, trying to upload a file that
exceeds this limit results in the following error being returned to the
client:

{"error":{"code":null,"message":"MultipartError { source: failed to read stream }","param":null,"type":"internal_server_error"}}

The same error is also logged in the ai-router log.

As OpenAI supports uploading audio files much larger than 2MB, we need
to increase this limit. While the OpenAI API supports files up to 25MB,
axum sets the default limit relatively low for security reasons, so
let's keep this value by default, but allow the ai-router admin to
override this limit via a new config option.

To improve the error message, we use tower-http's RequestBodyLimitLayer
instead of axum's DefaultBodyLimit. With this, when the limit is hit,
the error returned to the client looks like this:

length limit exceeded

While this deviates from the OpenAI compatible JSON errors we use, it
better indicates what is actually going on. Unfortunately it seems
impossible to customize this response without creating our own layer and
service to handle this, so for now this will have to do.