examples: add template for mistral chat completions
This template is based on the template for Mistral-7B-Instruct-v0.2 on
the Huggingface hub:
https://huggingface.co/mistralai/Mistral-7B-Instruct-v0.2/blob/main/tokenizer_config.json

The variables used in the original template have been added to the Jinja
file to make the actual template work. The template has been broken up
in multiple lines to improve readability. It also supports a system
prompt, if it is the first message sent by the client, by injecting it
before the first user message.