examples: add example template to disable completions
The ai-router config currently doesn't distinguish between chat and
legacy completion models; if a chat_completion model is configured, it
can be used for both chat and legacy completions. If for some reason you
want to disallow legacy completions for a model, this example template
can be used to do so.