errors: add generic BadRequestError
We're currently returning an InternalServerError, with HTTP status code
500 for most errors. When the OpenAI python library receives an HTTP 500
status code, it will retry the request twice by default. This is
suboptimal if we know the request will fail the next time.

Add a generic BadRequestError that returns HTTP status code 400. When
the OpenAI python library receives and HTTP 400, it will not retry.

Finally refactor InputExceededError to use the new error type.