startup: add SIGTERM signal handler
When running containerized as PID 1, we're receiving SIGTERM when the
container is being stopped. As we're currently ignoring SIGTERM, the
ai-router process appears to hang until a SIGKILL signal is received.
This slows down stopping the container and potentially even restarting a
system running ai-router.

Add a SIGTERM handler to avoid this.

Reported-by: Nick Bento <nikku.b@gmail.com>